Fix "Specified cast is not valid." error from users.expiry reads
Errors like the following were happening for all insers, updates
and selects from the user table (e.g. automatic greylisting,
broadcast commands, and manual list commands).

2017-10-10 23:54:09,177 [Main] ERROR - GlobalIntel failed
System.InvalidCastException: Specified cast is not valid.
  at Mono.Data.Sqlite.SqliteDataReader.VerifyType (System.Int32 i, System.Data.DbType typ)
  at Mono.Data.Sqlite.SqliteDataReader.GetString (System.Int32 i)
  at CVNBot.ListManager.GlobalIntel (System.String username)

The db schema declares users.expiry as integer(32).

As of Mono 4, it seems, SqliteDataReader#GetString is now strict
about only allowing it to be called if the underlying field is in
fact a string. Given it is of type integer, we should be calling
GetInt32 instead.

Changes:
* Fix all reading of expiry field to use GetInt32
* Update parseExpiryDate() to accept int instead of string

This then causes:

[Main] ERROR - Failed to handle broadcast command
System.ArgumentOutOfRangeException: Ticks must be between DateTime.MinValue.Ticks and DateTime.MaxValue.Ticks.
Parameter name: ticks
  at System.DateTime..ctor (System.Int64 ticks)
  at CVNBot.ListManager.parseExpiryDate (System.Int32 expiry)
  at CVNBot.ListManager.showUserOnList (System.String username, System.String project)
  at CVNBot.ListManager.addUserToList
  at CVNBot.Program.irc_OnChannelNotice

Looks like retreiving the field from the sqlite database as Int32 makes it
roll over because the ticks values in C# are actually much bigger and need Int64.
Casting afterwards just produces an invalid negative value due to rolling over.
Storing it in SQLite seems fine, and previous use of it as string was also
fine but using it as Int32 started breaking things.

Changes:
* Fix all reading of expiry field to use GetInt64
* Update parseExpiryDate() to accept long, not int or string