Fix "Specified cast is not valid." error from users.type reads
Errors like the following were happening for all inserts:

2017-10-11 00:17:01,775 [Main] ERROR - Error in handleListCommand
System.InvalidCastException: Specified cast is not valid.
  at Mono.Data.Sqlite.SqliteDataReader.VerifyType (System.Int32 i, System.Data.DbType typ)
  at Mono.Data.Sqlite.SqliteDataReader.GetString (System.Int32 i)
  at CVNBot.ListManager.classifyEditor (System.String username, System.String project)
  at CVNBot.ListManager.addUserToList
  at CVNBot.ListManager.handleListCommand

The db schema declares users.type as integer(32).

As of Mono 4, it seems, SqliteDataReader#GetString is now strict
about only allowing it to be called if the underlying field is in
fact a string. Given it is of type integer, we should be calling
GetInt32 instead.