Logger: Include Nick in syslog messages
This makes it easier to figure out which messages are coming from
when running multiple instances.

In theory one can look for the line that loads the ini settings
and then remember the PID, but it's tricky (especially when the
instances run for several days and logs have rolled over).

Also makes it easier to figure out what's happening in case a
bot is stuck in a shutdown/restart loop, in which case the PID
will change constantly.