LICENSE: Update copyright statement
Give the main contributors direct credit in the license for better
visibility (the AUTHORS.txt file is optional for distribution,
and wasn't maintainers very well, either).

Also add a mailmap file to normalise multiple names/addresses in
the commit history that relate to the same person, used by
the git-log and git-shortlog commands.

Also rename MIT-LICENSE.txt to LICENSE.txt for simplicity.