build: Use newer VS2017 format for csproj file
This change removes a lot of project organisation information
that doesn't seem to be needed or relevant for CVNBot.

Inspired by <https://natemcmaster.com/blog/2017/03/09/vs2015-to-vs2017-upgrade/>.

Many properties in csproj were the same as the defaults (not needed).
For example, 'Configuration' Debug/Release with Debug/optimize etc. exists
by default in MSBuild. No need to specify.

Also, we had many properties that were not used (e.g. Update*, Install*,
Publish*, etc.).

By using the VS2017-style format of 'csproj' format, and the new SDK,
we get the following benefits:

* Automatic creation of AssemblyInfo.
* Automatic inclusion of **/*.cs files.

This new format is not supported by xbuild (deprecated), we must
switch to MSBuild instead. This also requires support to be raised
to Mono 5+, because Mono 4 does not include MSBuild by default.

The new MSBuild needs `obj/project.assets.json` to exist before the
"Build" step can start. This file can be created automatically with
the "Restore" step, like so:

    msbuild src/CVNBot.sln /t:Restore /p:Configuration=Release

And then the regular command will work (default is /t:Clean;Build)

    msbuild src/CVNBot.sln /p:Configuration=Release

To keep it simple and not need two commands, I have made "restore"
part of the default target in CVNBot.csproj.