build: Support VS2019 and MSBuild 16 (#61)
The newer MSBuild has a new feature that automatically creates
a CVNBot.exe.config file in the bin/Release or bin/Debug directory
at the end of the build process:

> $ msbuild src/CVNBot.sln /p:Configuration=Debug
> …
> _CopyOutOfDateSourceItemsToOutputDirectory:
>   Copying "CVNBot-sample.ini" to "CVNBot/bin/Debug/CVNBot.ini".
>   Copying "CVNBot.exe.debug.config" to "CVNBot/bin/Debug/CVNBot.exe.config".
>   …
> _CopyAppConfigFile:
>   Copying "CVNBot/obj/Debug/CVNBot.exe.withSupportedRuntime.config"
>   to "CVNBot/bin/> Debug/CVNBot.exe.config".

The last one is the problem. It overwrites our `CVNBot.exe.config`
file with a different file. Thus there is no more console log or
syslog appender.

Disable this new feature by setting <GenerateSupportedRuntime> to false.

Also, MSBuild16 recommends adding <startup> to all config files, which it
does by default in its override, so I'm adding that as well.